+++ frontend/src/pages/Update/UpdSector.jsx
@@ -1,10 +1,13 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { CommonInput } from "../../components/CommonInput/CommonInput";
 import { CommonButton } from "../../components/CommonButton/CommonButton";
 import { UsersFour } from "@phosphor-icons/react";
+import { updateSector } from "../../functions/sectorManagement";
 
-export const UpdSector = ({ sector, updateSector }) => {
+export const UpdSector = ({ sector, navigate }) => {
+  const [onLoadState, setOnLoadState] = useState(false);
+  
   const {
     register,
     setValue,
@@ -17,14 +20,23 @@ export const UpdSector = ({ sector, updateSector }) => {
     setValue("id", `${sector.id}`);
   }, []);
 
+  const updateData = async (data) => {
+    setOnLoadState(true);
+
+    const response = await updateSector(data);
+    if (response) {
+      setTimeout(navigate("/sector"), 1000);
+      return;
+    }
+
+    setOnLoadState(false);
+  };
+
   return (
     <div className="flex flex-col gap-5 w-full">
       <h1 className="text-xl">Atualizar Setor</h1>
 
-      <form
-        className="flex flex-col gap-5"
-        onSubmit={handleSubmit(updateSector)}
-      >
+      <form className="flex flex-col gap-5" onSubmit={handleSubmit(updateData)}>
         <div className="flex flex-col gap-5">
           <label htmlFor="sectorName">Nome</label>
           <CommonInput
@@ -43,6 +55,7 @@ export const UpdSector = ({ sector, updateSector }) => {
         <CommonButton
           icon={<UsersFour size={24} />}
           content="Atualizar setor"
+          onLoadState={onLoadState}
         />
       </form>
     </div>
