+++ frontend/src/routes/Update/UpdateUser.jsx
@@ -2,10 +2,9 @@ import { LoadingComponent } from "../../components/LoadingComponent/LoadingCompo
 import { useNavigate, useParams } from "react-router-dom";
 import { useContext, useEffect, useState } from "react";
 import { Container } from "../../components/Container";
-import { getUser, updateUser } from "../../functions/userManagement";
+import { getUser } from "../../functions/userManagement";
 import { UpdUser } from "../../pages/Update/UpdUser";
 import { AuthenticationContext } from "../../provider/AuthenticationProvider";
-import { updateLiveUserData } from "../../functions/auth";
 
 export const UpdateUser = () => {
   const navigate = useNavigate();
@@ -42,23 +41,10 @@ export const UpdateUser = () => {
     getData();
   }, []);
 
-  const updateData = async (data) => {
-    const response = await updateUser(data);
-    if (response) {
-      if (id) {
-        setTimeout(navigate("/user"), 1000);
-      } else {
-  
-        updateLiveUserData(response, setUserData);
-        setTimeout(navigate("/about"), 1000);
-      }
-    }
-  };
-
   return (
     <LoadingComponent isLoading={isLoading}>
       <Container>
-        <UpdUser user={user} updateUser={updateData} />
+        <UpdUser user={user} navigate={navigate}/>
       </Container>
     </LoadingComponent>
   );
